Lesson6 внес в проект то, что сделали на уроке 3 с отображением списоков, перенес на спрингбут, собираюсь прикручивать энтитименеджер бд

+++ src/main/java/ru/gb/spring1lesson6/model/Product.java
@@ -12,7 +12,8 @@ import lombok.*;
 public class Product {
     private Integer id;
     private String name;
-    private double cost;
+    private Integer cost;
+
 
     @Override
     public boolean equals(Object o) {
